% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional.R
\docType{data}
\name{none}
\alias{none}
\title{None}
\format{An object of class \code{optional} of length 1.}
\usage{
none
}
\description{
Indicates an invalid variable.
Might be returned by an optional function 
(see \code{?make_opt()})
}
\examples{
a <- none
a
## [1] None
}
\seealso{
option(), opt_unwrap()
}
\keyword{datasets}
