% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional.R
\name{some}
\alias{some}
\title{Some
Make a variable optional

Note that `some(some(i)) == some(i)` 
and `some(none) == FALSE`}
\usage{
some(arg)
}
\arguments{
\item{arg}{The variable to make optional}
}
\value{
`some(arg)`
}
\examples{
a <- some(5)
class(a)
> [1] "optional"
a == 5
> [1] TRUE
}
\seealso{
`none`, `opt_unwrap`
}

