% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional.R
\name{some}
\alias{some}
\title{some}
\usage{
some(arg)
}
\arguments{
\item{arg}{The variable to check existence}
}
\value{
TRUE if \code{arg} is an optional variable 
and if it is not none, else returns FALSE
}
\description{
Check if a optional object equals none
}
\seealso{
option(), none
a <- option(1)
some(a)
## [1] TRUE
b <- none
some(b)
## [1] FALSE
}
