% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional.R
\name{some}
\alias{some}
\title{Some}
\usage{
some(arg)
}
\arguments{
\item{arg}{The variable to make optional}
}
\value{
\code{arg} as \code{optional}
}
\description{
Make a variable optional. 

\code{optional} is an object wrapper which indicates
whether the object is valid or not.
}
\details{
Note that \code{some(some(i)) == some(i)}
and \code{some(none) == FALSE}

Operators and print will have the same behavior with 
an optional than with its base type.
}
\examples{
a <- some(5)
class(a)
## [1] "optional"

a == 5
## [1] TRUE

a
## [1] 5
}
\seealso{
none, opt_unwrap(), make_opt()
}

